+++ NEWS.md
@@ -7,6 +7,7 @@ title: "NEWS for the emmeans package"
     We now suppress it when all the prior weights are equal.
   * Fix to `MuMIn` support with `subset` argument (#455)
   * Repair to coding error for nested models (#457)
+  * Added `glmtoolbox::glmgee` support (#454)
     
 
 ## emmeans 1.8.9